machine_learning/requirements.txt
scikit-learn==1.0.2
pandas==1.4.2
numpy==1.22.3
joblib==1.1.0
kafka-python==2.0.2
matplotlib==3.5.1
seaborn==0.11.2
